data-eng-proj-main/requirements.txt
faker
pandas
kafka-python
